requirements.txt
aiohttp==3.6.0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.9
discord-py-slash-command==3.0.3
discord.py==1.7.3
docopt==0.6.2
frozenlist==1.2.0
idna==2.5
multidict==4.0
requests==2.26.0
SQLAlchemy==1.3.3
twitter==1.18.0
typing_extensions==4.0.1
urllib3==1.26.7
websockets==10.1
yarg==0.1.9
yarl==1.7.2
